mortgagecalculator.py
This commit is the first version of the GUI program. We have a basic GUI with user inputs for loan amount, loan term, and interest rate. This will calculate the monthly payment. There are also buttons that allow the user to either clear the entry fields or exit the program altogether. The next step is to add more user entry fields for HOA fees, property tax, and home insurance while also upgrading the UI design of the application.